+++ src/meta/api/src/schema_api_impl.rs
@@ -77,27 +77,27 @@ use common_meta_app::share::ShareGrantObject;
 use common_meta_app::share::ShareGrantObjectPrivilege;
 use common_meta_app::share::ShareId;
 use common_meta_app::share::ShareNameIdent;
-use common_meta_types::app_error::AppError;
-use common_meta_types::app_error::CreateDatabaseWithDropTime;
-use common_meta_types::app_error::CreateTableWithDropTime;
-use common_meta_types::app_error::DatabaseAlreadyExists;
-use common_meta_types::app_error::DropDbWithDropTime;
-use common_meta_types::app_error::DropTableWithDropTime;
-use common_meta_types::app_error::ShareHasNoGrantedDatabase;
-use common_meta_types::app_error::ShareHasNoGrantedPrivilege;
-use common_meta_types::app_error::TableAlreadyExists;
-use common_meta_types::app_error::TableVersionMismatched;
-use common_meta_types::app_error::TxnRetryMaxTimes;
-use common_meta_types::app_error::UndropDbHasNoHistory;
-use common_meta_types::app_error::UndropDbWithNoDropTime;
-use common_meta_types::app_error::UndropTableAlreadyExists;
-use common_meta_types::app_error::UndropTableHasNoHistory;
-use common_meta_types::app_error::UndropTableWithNoDropTime;
-use common_meta_types::app_error::UnknownShareAccounts;
-use common_meta_types::app_error::UnknownTable;
-use common_meta_types::app_error::UnknownTableId;
-use common_meta_types::app_error::WrongShare;
-use common_meta_types::app_error::WrongShareObject;
+use common_meta_types::errors::app_error::AppError;
+use common_meta_types::errors::app_error::CreateDatabaseWithDropTime;
+use common_meta_types::errors::app_error::CreateTableWithDropTime;
+use common_meta_types::errors::app_error::DatabaseAlreadyExists;
+use common_meta_types::errors::app_error::DropDbWithDropTime;
+use common_meta_types::errors::app_error::DropTableWithDropTime;
+use common_meta_types::errors::app_error::ShareHasNoGrantedDatabase;
+use common_meta_types::errors::app_error::ShareHasNoGrantedPrivilege;
+use common_meta_types::errors::app_error::TableAlreadyExists;
+use common_meta_types::errors::app_error::TableVersionMismatched;
+use common_meta_types::errors::app_error::TxnRetryMaxTimes;
+use common_meta_types::errors::app_error::UndropDbHasNoHistory;
+use common_meta_types::errors::app_error::UndropDbWithNoDropTime;
+use common_meta_types::errors::app_error::UndropTableAlreadyExists;
+use common_meta_types::errors::app_error::UndropTableHasNoHistory;
+use common_meta_types::errors::app_error::UndropTableWithNoDropTime;
+use common_meta_types::errors::app_error::UnknownShareAccounts;
+use common_meta_types::errors::app_error::UnknownTable;
+use common_meta_types::errors::app_error::UnknownTableId;
+use common_meta_types::errors::app_error::WrongShare;
+use common_meta_types::errors::app_error::WrongShareObject;
 use common_meta_types::ConditionResult;
 use common_meta_types::GCDroppedDataReply;
 use common_meta_types::GCDroppedDataReq;
