+++ src/meta/api/src/share_api_impl.rs
@@ -21,15 +21,15 @@ use common_meta_app::schema::TableIdToName;
 use common_meta_app::schema::TableMeta;
 use common_meta_app::schema::TableNameIdent;
 use common_meta_app::share::*;
-use common_meta_types::app_error::AppError;
-use common_meta_types::app_error::ShareAccountsAlreadyExists;
-use common_meta_types::app_error::ShareAlreadyExists;
-use common_meta_types::app_error::TxnRetryMaxTimes;
-use common_meta_types::app_error::UnknownShare;
-use common_meta_types::app_error::UnknownShareAccounts;
-use common_meta_types::app_error::UnknownTable;
-use common_meta_types::app_error::WrongShare;
-use common_meta_types::app_error::WrongShareObject;
+use common_meta_types::errors::app_error::AppError;
+use common_meta_types::errors::app_error::ShareAccountsAlreadyExists;
+use common_meta_types::errors::app_error::ShareAlreadyExists;
+use common_meta_types::errors::app_error::TxnRetryMaxTimes;
+use common_meta_types::errors::app_error::UnknownShare;
+use common_meta_types::errors::app_error::UnknownShareAccounts;
+use common_meta_types::errors::app_error::UnknownTable;
+use common_meta_types::errors::app_error::WrongShare;
+use common_meta_types::errors::app_error::WrongShareObject;
 use common_meta_types::ConditionResult::Eq;
 use common_meta_types::KVAppError;
 use common_meta_types::TxnCondition;
