+++ src/meta/api/src/util.rs
@@ -20,13 +20,13 @@ use common_meta_app::schema::DatabaseMeta;
 use common_meta_app::schema::DatabaseNameIdent;
 use common_meta_app::schema::TableNameIdent;
 use common_meta_app::share::*;
-use common_meta_types::app_error::AppError;
-use common_meta_types::app_error::ShareHasNoGrantedDatabase;
-use common_meta_types::app_error::UnknownDatabase;
-use common_meta_types::app_error::UnknownShare;
-use common_meta_types::app_error::UnknownShareAccounts;
-use common_meta_types::app_error::UnknownShareId;
-use common_meta_types::app_error::UnknownTable;
+use common_meta_types::errors::app_error::AppError;
+use common_meta_types::errors::app_error::ShareHasNoGrantedDatabase;
+use common_meta_types::errors::app_error::UnknownDatabase;
+use common_meta_types::errors::app_error::UnknownShare;
+use common_meta_types::errors::app_error::UnknownShareAccounts;
+use common_meta_types::errors::app_error::UnknownShareId;
+use common_meta_types::errors::app_error::UnknownTable;
 use common_meta_types::txn_condition::Target;
 use common_meta_types::txn_op::Request;
 use common_meta_types::ConditionResult;
