+++ src/meta/app/src/share/share.rs
@@ -20,8 +20,8 @@ use std::fmt::Formatter;
 
 use common_datavalues::chrono::DateTime;
 use common_datavalues::chrono::Utc;
-use common_meta_types::app_error::AppError;
-use common_meta_types::app_error::WrongShareObject;
+use common_meta_types::errors::app_error::AppError;
+use common_meta_types::errors::app_error::WrongShareObject;
 use common_meta_types::KVAppError;
 use enumflags2::bitflags;
 use enumflags2::BitFlags;
