+++ src/meta/types/src/errors/meta_bytes_error.rs
