+++ src/meta/types/src/lib.rs
@@ -22,28 +22,16 @@ mod cluster;
 mod cmd;
 pub mod config;
 mod endpoint;
-mod errors;
-mod kv_app_errors;
+pub mod errors;
 mod kv_message;
 mod log_entry;
 mod match_seq;
 mod message;
-mod meta_bytes_error;
-mod meta_client_errors;
-mod meta_errors;
-mod meta_errors_into;
-mod meta_forward_errors;
-mod meta_handshake_errors;
-mod meta_management_error;
-mod meta_network_errors;
-mod meta_raft_errors;
-mod meta_startup_errors;
-mod meta_storage_errors;
 mod operation;
 mod raft_txid;
 mod raft_types;
 mod role_info;
-mod rpc_errors;
+mod seq_errors;
 mod seq_num;
 mod seq_value;
 mod tenant_quota;
@@ -58,7 +46,6 @@ mod user_setting;
 mod user_stage;
 mod with;
 
-pub mod app_error;
 pub mod error_context;
 mod principal_identity;
 mod proto_display;
@@ -74,26 +61,6 @@ pub mod protobuf {
 
     pub const FILE_DESCRIPTOR_SET: &[u8] = tonic::include_file_descriptor_set!("meta_descriptor");
 }
-pub use app_error::AppError;
-pub use app_error::CreateDatabaseWithDropTime;
-pub use app_error::CreateTableWithDropTime;
-pub use app_error::DatabaseAlreadyExists;
-pub use app_error::DropDbWithDropTime;
-pub use app_error::DropTableWithDropTime;
-pub use app_error::ShareAlreadyExists;
-pub use app_error::TableAlreadyExists;
-pub use app_error::TableVersionMismatched;
-pub use app_error::UndropDbHasNoHistory;
-pub use app_error::UndropDbWithNoDropTime;
-pub use app_error::UndropTableAlreadyExists;
-pub use app_error::UndropTableHasNoHistory;
-pub use app_error::UndropTableWithNoDropTime;
-pub use app_error::UnknownDatabase;
-pub use app_error::UnknownDatabaseId;
-pub use app_error::UnknownShare;
-pub use app_error::UnknownTable;
-pub use app_error::UnknownTableId;
-pub use app_error::WrongShareObject;
 pub use applied_state::AppliedState;
 pub use change::Change;
 pub use cluster::Node;
@@ -102,8 +69,53 @@ pub use cluster::Slot;
 pub use cmd::Cmd;
 pub use cmd::UpsertKV;
 pub use endpoint::Endpoint;
-pub use errors::ConflictSeq;
-pub use kv_app_errors::KVAppError;
+pub use errors::app_error::AppError;
+pub use errors::app_error::CreateDatabaseWithDropTime;
+pub use errors::app_error::CreateTableWithDropTime;
+pub use errors::app_error::DatabaseAlreadyExists;
+pub use errors::app_error::DropDbWithDropTime;
+pub use errors::app_error::DropTableWithDropTime;
+pub use errors::app_error::ShareAlreadyExists;
+pub use errors::app_error::TableAlreadyExists;
+pub use errors::app_error::TableVersionMismatched;
+pub use errors::app_error::UndropDbHasNoHistory;
+pub use errors::app_error::UndropDbWithNoDropTime;
+pub use errors::app_error::UndropTableAlreadyExists;
+pub use errors::app_error::UndropTableHasNoHistory;
+pub use errors::app_error::UndropTableWithNoDropTime;
+pub use errors::app_error::UnknownDatabase;
+pub use errors::app_error::UnknownDatabaseId;
+pub use errors::app_error::UnknownShare;
+pub use errors::app_error::UnknownTable;
+pub use errors::app_error::UnknownTableId;
+pub use errors::app_error::WrongShareObject;
+pub use errors::kv_app_errors::KVAppError;
+pub use errors::meta_api_errors::MetaAPIError;
+pub use errors::meta_api_errors::MetaDataError;
+pub use errors::meta_api_errors::MetaDataReadError;
+pub use errors::meta_api_errors::MetaOperationError;
+pub use errors::meta_bytes_error::MetaBytesError;
+pub use errors::meta_client_errors::MetaClientError;
+pub use errors::meta_errors::MetaError;
+pub use errors::meta_errors::MetaResult;
+pub use errors::meta_handshake_errors::MetaHandshakeError;
+pub use errors::meta_management_error::MetaManagementError;
+pub use errors::meta_network_errors::ConnectionError;
+pub use errors::meta_network_errors::InvalidArgument;
+pub use errors::meta_network_errors::InvalidReply;
+pub use errors::meta_network_errors::MetaNetworkError;
+pub use errors::meta_network_errors::MetaNetworkResult;
+pub use errors::meta_raft_errors::ChangeMembershipError;
+pub use errors::meta_raft_errors::Fatal;
+pub use errors::meta_raft_errors::ForwardToLeader;
+pub use errors::meta_raft_errors::InitializeError;
+pub use errors::meta_raft_errors::RaftChangeMembershipError;
+pub use errors::meta_raft_errors::RaftWriteError;
+pub use errors::meta_raft_errors::RetryableError;
+pub use errors::meta_startup_errors::MetaStartupError;
+pub use errors::meta_storage_errors::MetaStorageError;
+pub use errors::meta_storage_errors::MetaStorageResult;
+pub use errors::rpc_errors::ForwardRPCError;
 pub use kv_message::GetKVReply;
 pub use kv_message::GetKVReq;
 pub use kv_message::ListKVReply;
@@ -120,31 +132,6 @@ pub use message::ForwardRequestBody;
 pub use message::ForwardResponse;
 pub use message::JoinRequest;
 pub use message::LeaveRequest;
-pub use meta_bytes_error::MetaBytesError;
-pub use meta_client_errors::MetaClientError;
-pub use meta_errors::MetaError;
-pub use meta_errors::MetaResult;
-pub use meta_forward_errors::MetaAPIError;
-pub use meta_forward_errors::MetaDataError;
-pub use meta_forward_errors::MetaDataReadError;
-pub use meta_forward_errors::MetaOperationError;
-pub use meta_handshake_errors::MetaHandshakeError;
-pub use meta_management_error::MetaManagementError;
-pub use meta_network_errors::ConnectionError;
-pub use meta_network_errors::InvalidArgument;
-pub use meta_network_errors::InvalidReply;
-pub use meta_network_errors::MetaNetworkError;
-pub use meta_network_errors::MetaNetworkResult;
-pub use meta_raft_errors::ChangeMembershipError;
-pub use meta_raft_errors::Fatal;
-pub use meta_raft_errors::ForwardToLeader;
-pub use meta_raft_errors::InitializeError;
-pub use meta_raft_errors::RaftChangeMembershipError;
-pub use meta_raft_errors::RaftWriteError;
-pub use meta_raft_errors::RetryableError;
-pub use meta_startup_errors::MetaStartupError;
-pub use meta_storage_errors::MetaStorageError;
-pub use meta_storage_errors::MetaStorageResult;
 pub use operation::GCDroppedDataReply;
 pub use operation::GCDroppedDataReq;
 pub use operation::MetaId;
@@ -174,7 +161,7 @@ pub use raft_types::NodeId;
 pub use raft_types::Term;
 pub use role_info::RoleInfo;
 pub use role_info::RoleInfoSerdeError;
-pub use rpc_errors::ForwardRPCError;
+pub use seq_errors::ConflictSeq;
 pub use seq_num::SeqNum;
 pub use seq_value::IntoSeqV;
 pub use seq_value::KVMeta;
